fix: emit copy workflow only only when file isn't growing
It could happen that the file is growing on the source storage when
checkAndEmitCopyWorkflow is called. It is incorrect to start the copy
before the file is fully present on source storage. The solution is to
probe the file size twice, 3 seconds apart, and emit the copy workflow
only when the size hasn't changed. If it has changed, just wait for an
add/change event.